Streamlined non-integrated IntApp connections
This seems to fix #325

+++ mRemoteV1/Connection/Protocol/IntegratedProgram.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Drawing;
 using System.Threading;
 using System.Windows.Forms;
+using mRemoteNG.Messages;
 
 
 namespace mRemoteNG.Connection.Protocol
@@ -32,11 +33,18 @@ namespace mRemoteNG.Connection.Protocol
 		{
 			try
 			{
-				if (_externalTool.TryIntegrate == false)
+                Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, $"Attempting to start: {_externalTool.DisplayName}", true);
+
+                if (_externalTool.TryIntegrate == false)
 				{
 					_externalTool.Start(InterfaceControl.Info);
-					Close();
-					return false;
+                    /* Don't call close here... There's nothing for the override to do in this case since 
+                     * _process is not created in this scenario. When returning false, ProtocolBase.Close()
+                     * will be called - which is just going to call IntegratedProgram.Close() again anyway...
+                     * Close();
+                     */
+                    Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, $"Assuming no other errors/exceptions occurred immediately before this message regarding {_externalTool.DisplayName}, the next \"closed by user\" message can be ignored", true);
+                    return false;
 				}
 
                 var argParser = new ExternalToolArgumentParser(_externalTool.ConnectionInfo);
@@ -72,10 +80,10 @@ namespace mRemoteNG.Connection.Protocol
 				}
 						
 				NativeMethods.SetParent(_handle, InterfaceControl.Handle);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, Language.strIntAppStuff, true);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, string.Format(Language.strIntAppHandle, _handle), true);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, string.Format(Language.strIntAppTitle, _process.MainWindowTitle), true);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, string.Format(Language.strIntAppParentHandle, InterfaceControl.Parent.Handle), true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, Language.strIntAppStuff, true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, string.Format(Language.strIntAppHandle, _handle), true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, string.Format(Language.strIntAppTitle, _process.MainWindowTitle), true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, string.Format(Language.strIntAppParentHandle, InterfaceControl.Parent.Handle), true);
 						
 				Resize(this, new EventArgs());
 				base.Connect();
@@ -116,31 +124,38 @@ namespace mRemoteNG.Connection.Protocol
 				
 		public override void Close()
 		{
-			try
-			{
-				if (!_process.HasExited)
-				{
-					_process.Kill();
-				}
-			}
-			catch (Exception ex)
-			{
-				Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppKillFailed, ex);
-			}
-					
-			try
-			{
-				if (!_process.HasExited)
-				{
-					_process.Dispose();
-				}
-			}
-			catch (Exception ex)
-			{
-				Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppDisposeFailed, ex);
-			}
-					
-			base.Close();
+            /* only attempt this if we have a valid process object
+             * Non-integated tools will still call base.Close() and don't have a valid process object.
+             * See Connect() above... This just muddies up the log.
+             */
+            if (_process != null)
+		    {
+		        try
+		        {
+		            if (!_process.HasExited)
+		            {
+		                _process.Kill();
+		            }
+		        }
+		        catch (Exception ex)
+		        {
+		            Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppKillFailed, ex);
+		        }
+
+		        try
+		        {
+		            if (!_process.HasExited)
+		            {
+		                _process.Dispose();
+		            }
+		        }
+		        catch (Exception ex)
+		        {
+		            Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppDisposeFailed, ex);
+		        }
+		    }
+
+		    base.Close();
 		}
         #endregion
